Refactored format to be consistent (and cleaner)
Still more work needs to be done on the query_id_from_email function. It should return a user object (or possibly an array of objects if multiple users have the same email).

The biggest issue currently is that the name implies that it returns an id, however, the code after it implies that it actually returns some sort of array or object. The behavior of the function should be consistent with it's name.